requirements.txt
spidev
dbus-python
pygobject